refactor: split ConversationViewModel into multiple smaller ViewModels
