+++ app/src/main/kotlin/com/wire/android/ui/home/conversations/call/ConversationCallViewState.kt
@@ -1,0 +1,6 @@
+package com.wire.android.ui.home.conversations.call
+
+data class ConversationCallViewState(
+    val hasOngoingCall: Boolean = false,
+    val hasEstablishedCall: Boolean = false
+)
